Add error handling to occurrence enrolling
